move ScrollableRegion to its own internal component
